kubernetes-shell/src/requirements.txt
urllib3==1.23
cloudshell-shell-core>=4.0.0,<4.1.0
cloudshell-cp-core>=1.0.0,<1.1.0
kubernetes==7.0.0
typing==3.6.6